refactor: extract core of cache to reusable composable
